Refactored project structure to have a valid project name (packages can't contain hyphens and should be lower case (!).
